refactor: Rename GraphQL processor to Mediajel processor
